Makes projects Net 5 again, adds AddRecipeToList, move code
* Makes projects .Net 5 again, so the plugin can be loaded in a existing NSSL server
* Moved dto into own files
* Made the server ipv4/ipv6 choose dynamically
