+++ NSSLServer.Core/NSSLServer.Core.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
     <PackageId>NSSLServer.Core</PackageId>
     <description>Core dll for creating a nssl server plugin.</description>
   </PropertyGroup>
