+++ NSSLServer.Database/NSSLServer.Database.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
     <PackageId>NSSLServer.Database</PackageId>
     <description>For creating a nssl server plugin, which uses the database.</description>
   </PropertyGroup>
