+++ NSSLServer.Plugin.Recipes/AssemblyAttributes.cs
@@ -1,0 +1,4 @@
+﻿
+using NSSLServer.Core.Extension;
+
+[assembly: Plugin()]
