+++ NSSLServer.Plugin.Shoppinglist/NSSLServer.Plugin.Shoppinglist.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
     <PackageId>NSSLServer.Plugin.Shoppinglist</PackageId>
     <description>Extension Plugin for nssl server for shoppinglist functionality.</description>
   </PropertyGroup>
